requirements.txt
google-adk
mcp
fastmcp
python-dotenv
httpx
google-auth
google-api-python-client
starlette